# api-gateway-user/docs/swagger.yaml
basePath: /
definitions:
...
  genproto.Pagination:
    properties:
      limit:
        type: integer
      offset:
        type: integer
    type: object
  genproto.PostCreate:
    properties:
      PictureUrls:
        items:
          type: string
        type: array
      content:
        type: string
      title:
        type: string
    type: object
  genproto.PostGet:
    properties:
      CreatedAt:
        type: string
      PictureUrls:
        items:
          type: string
        type: array
      content:
        type: string
      id:
        type: string
      title:
        type: string
    type: object
  genproto.PostList:
    properties:
      Pagination:
        $ref: '#/definitions/genproto.Pagination'
      Posts:
        items:
          $ref: '#/definitions/genproto.PostGet'
        type: array
      TotalCount:
        type: integer
    type: object
  genproto.PostPicture:
    properties:
      PictureUrl:
        type: string
      PostId:
        type: string
    type: object
  genproto.PostUpdate:
    properties:
      Body:
        $ref: '#/definitions/genproto.PostUpt'
      id:
        type: string
    type: object
  genproto.PostUpt:
    properties:
      Content:
        type: string
      Title:
        type: string
    type: object
...
info:
  contact: {}
  description: API for Instant Delivery resources
  title: Finance Tracker API Documentation
  version: "1.0"
paths:
...
  /posts/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a post by its ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Post details
          schema:
            $ref: '#/definitions/genproto.PostGet'
        "400":
          description: Invalid request
          schema:
            type: string
        "404":
          description: Post not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Post
      tags:
      - Post
  /posts/create:
    post:
      consumes:
      - application/json
      description: Create a new post
      parameters:
      - description: Post data
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/genproto.PostCreate'
      produces:
      - application/json
      responses:
        "200":
          description: Post created successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create Post
      tags:
      - Post
  /posts/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a post by its ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Post deleted successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "404":
          description: Post not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete Post
      tags:
      - Post
  /posts/list:
    get:
      consumes:
      - application/json
      description: Retrieve a list of posts with pagination
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of posts
          schema:
            $ref: '#/definitions/genproto.PostList'
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Posts
      tags:
      - Post
  /posts/picture/add:
    post:
      consumes:
      - application/json
      description: Add a picture to a post
      parameters:
      - description: Post picture data
        in: body
        name: picture
        required: true
        schema:
          $ref: '#/definitions/genproto.PostPicture'
      produces:
      - application/json
      responses:
        "200":
          description: Picture added successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Add Post Picture
      tags:
      - Post
  /posts/picture/delete:
    post:
      consumes:
      - application/json
      description: Delete a picture from a post
      parameters:
      - description: Post picture data
        in: body
        name: picture
        required: true
        schema:
          $ref: '#/definitions/genproto.PostPicture'
      produces:
      - application/json
      responses:
        "200":
          description: Picture deleted successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete Post Picture
      tags:
      - Post
  /posts/update/{id}:
    put:
      consumes:
      - application/json
      description: Update an existing post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      - description: Post data
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/genproto.PostUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Post updated successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update Post
      tags:
      - Post
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"